Static analysis improvements

+++ database/migrations/2017_10_04_045605_create_initial_roles_and_permissions.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 use Illuminate\Database\Migrations\Migration;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
@@ -8,10 +10,8 @@ class CreateInitialRolesAndPermissions extends Migration
 {
     /**
      * Run the migrations.
-     *
-     * @return void
      */
-    public function up()
+    public function up(): void
     {
         // Reset cached roles and permissions
         app()['cache']->forget('spatie.permission.cache');
@@ -226,14 +226,15 @@ class CreateInitialRolesAndPermissions extends Migration
 
     /**
      * Reverse the migrations.
-     *
-     * @return void
      */
-    public function down()
+    public function down(): void
     {
         $roles = ['admin', 'officer-i', 'officer-ii', 'core', 'member', 'non-member'];
         foreach ($roles as $role) {
             $dbRole = Role::where('name', $role)->first();
+            if (null === $dbRole) {
+                continue;
+            }
             $dbRole->delete();
         }
     }
